airserver/requirements.txt
# airhelper
PyYAML
CoAPthon3
pycryptodomex

# airserver
websockets
